app/requirements.txt
requests
python-dotenv
pika
schedule